Refactor application helper
