+++ app/helpers/application_helper.rb
@@ -41,27 +41,23 @@
 #
 module ApplicationHelper
   # All helpers are autoloaded under Zeitwerk
-  def safe_empty
-    "".html_safe
-  end
 
-  def safe_br
-    "<br/>".html_safe
-  end
-
-  def safe_nbsp
-    "&nbsp;".html_safe
-  end
-
-  def safe_spinner(text = "")
-    "#{text}<span class='spinner-right mx-2'></span>".html_safe
-  end
-
-  # Return escaped HTML.
-  #
-  #   "<i>X</i>"  -->  "&lt;i&gt;X&lt;/i&gt;"
-  def escape_html(html)
-    h(html.to_str)
+  def css_theme
+    if in_admin_mode?
+      "Admin"
+    elsif session[:real_user_id].present?
+      "Sudo"
+    elsif browser.bot? || !@user
+      MO.default_theme
+    elsif MO.themes.member?(controller.action_name)
+      # when looking at a theme's info page render it in that theme
+      controller.action_name
+    elsif @user && @user&.theme.present? &&
+          MO.themes.member?(@user.theme)
+      @user.theme
+    elsif @user
+      MO.themes.sample
+    end
   end
 
   # Return a width class for the layout content container.
@@ -82,38 +78,17 @@ module ApplicationHelper
     end
   end
 
-  def css_theme
-    if in_admin_mode?
-      "Admin"
-    elsif session[:real_user_id].present?
-      "Sudo"
-    elsif browser.bot? || !@user
-      MO.default_theme
-    elsif MO.themes.member?(controller.action_name)
-      # when looking at a theme's info page render it in that theme
-      controller.action_name
-    elsif @user && @user&.theme.present? &&
-          MO.themes.member?(@user.theme)
-      @user.theme
-    elsif @user
-      MO.themes.sample
-    end
-  end
-
   # --------- links and buttons ------------------------------------------------
 
-  # Call link_to with query params added.
-  def link_with_query(name = nil, options = nil, html_options = nil)
-    link_to(name, add_query_param(options), html_options)
-  end
-
-  # Take a query which can be coerced into a different model, and create a link
-  # to the results of that coerced query.  Return +nil+ if not coercable.
-  def link_to_coerced_query(query, model)
-    link = coerced_query_link(query, model)
-    return nil unless link
-
-    link_to(*link)
+  # contents of the <title> in html header
+  def title_tag_contents(action_name)
+    if @title.present?
+      @title.strip_html.html_safe
+    elsif TranslationString.where(tag: "title_for_#{action_name}").present?
+      :"title_for_#{action_name}".t
+    else
+      action_name.tr("_", " ").titleize
+    end
   end
 
   # link to next object in query results
@@ -146,77 +121,6 @@ module ApplicationHelper
     link_with_query("« #{:BACK.t}", path)
   end
 
-  # button to destroy object
-  # Used instead of link_to because method: :delete requires jquery_ujs library
-  # Sample usage:
-  #   destroy_button(target: article)
-  #   destroy_button(name: :destroy_object.t(type: :glossary_term),
-  #                  target: term)
-  #   destroy_button(
-  #     name: :destroy_object.t(type: :herbarium),
-  #     target: herbarium_path(@herbarium, back: url_after_delete(@herbarium))
-  #   )
-  def destroy_button(target:, name: :DESTROY.t, **args)
-    path = if target.is_a?(String)
-             target
-           else
-             add_query_param(send("#{target.type_tag}_path", target.id))
-           end
-    classes ||= "text-danger"
-    id ||= nil
-    unless target.is_a?(String)
-      classes += " destroy_#{target.type_tag}_link_#{target.id}"
-    end
-
-    html_options = {
-      method: :delete,
-      class: classes,
-      id: id,
-      data: { confirm: :are_you_sure.t }
-    }.merge(args)
-
-    button_to(name, path, html_options)
-  end
-
-  # POST to a path; used instead of a link because POST link requires js
-  # post_button(name: herbarium.name.t,
-  #             path: herbaria_merges_path(that: @merge.id,this: herbarium.id),
-  #             data: { confirm: :are_you_sure.t })
-  def post_button(name:, path:, **args)
-    html_options = {
-      method: :post,
-      class: ""
-    }.merge(args)
-
-    button_to(name, path, html_options)
-  end
-
-  # PUT to a path; used instead of a link because PUT link requires js
-  # put_button(name: herbarium.name.t,
-  #            path: herbarium_path(id: @herbarium.id),
-  #            data: { confirm: :are_you_sure.t })
-  def put_button(name:, path:, **args)
-    html_options = {
-      method: :put,
-      class: ""
-    }.merge(args)
-
-    button_to(name, path, html_options)
-  end
-
-  # PATCH to a path; used instead of a link because PATCH link requires js
-  # patch_button(name: herbarium.name.t,
-  #              path: herbarium_path(id: @herbarium.id),
-  #              data: { confirm: :are_you_sure.t })
-  def patch_button(name:, path:, **args)
-    html_options = {
-      method: :patch,
-      class: ""
-    }.merge(args)
-
-    button_to(name, path, html_options)
-  end
-
   # Convert @links in index views into a list of tabs for RHS tab set.
   def create_links(links)
     return [] unless links
@@ -231,38 +135,6 @@ module ApplicationHelper
 
   # ----------------------------------------------------------------------------
 
-  # Create an in-line white-space element approximately the given width in
-  # pixels.  It should be non-line-breakable, too.
-  def indent
-    content_tag(:span, "&nbsp;".html_safe, class: "ml-10px")
-  end
-
-  def spacer
-    content_tag(:span, "&nbsp;".html_safe, class: "mx-2")
-  end
-
-  def content_tag_if(condition, name, content_or_options_with_block = nil,
-                     options = nil, escape = true, &block)
-    return unless condition
-
-    content_tag(name, content_or_options_with_block, options, escape, &block)
-  end
-
-  def content_tag_unless(condition, name, content_or_options_with_block = nil,
-                         options = nil, escape = true, &block)
-    content_tag_if(!condition, name, content_or_options_with_block,
-                   options, escape, &block)
-  end
-
-  # Wrap an html object in '<span title="blah">' tag.  This has the effect of
-  # giving it context help (mouse-over popup) in most modern browsers.
-  #
-  #   <%= add_context_help(link, "Click here to do something.") %>
-  #
-  def add_context_help(object, help)
-    content_tag(:span, object, title: help, data: { toggle: "tooltip" })
-  end
-
   # Add something to the header from within view.  This can be called as many
   # times as necessary -- the application layout will mash them all together
   # and stick them at the end of the <tt>&gt;head&lt;/tt> section.
@@ -276,58 +148,6 @@ module ApplicationHelper
   def add_header(str)
     @header ||= safe_empty
     @header += str
-  end
-
-  # Create a table out of a list of Arrays.
-  #
-  #   make_table([[1,2],[3,4]])
-  #
-  # Produces:
-  #
-  #   <table>
-  #     <tr>
-  #       <td>1</td>
-  #       <td>2</td>
-  #     </tr>
-  #     <tr>
-  #       <td>3</td>
-  #       <td>4</td>
-  #     </tr>
-  #   </table>
-  #
-  def make_table(rows, table_opts = {}, tr_opts = {}, td_opts = {})
-    content_tag(:table, table_opts) do
-      rows.map do |row|
-        make_row(row, tr_opts, td_opts) + make_line(row, td_opts)
-      end.safe_join
-    end
-  end
-
-  def make_row(row, tr_opts = {}, td_opts = {})
-    content_tag(:tr, tr_opts) do
-      if row.is_a?(Array)
-        row.map do |cell|
-          make_cell(cell, td_opts)
-        end.safe_join
-      else
-        row
-      end
-    end
-  end
-
-  def make_cell(cell, td_opts = {})
-    content_tag(:td, cell.to_s, td_opts)
-  end
-
-  def make_line(_row, td_opts)
-    colspan = td_opts[:colspan]
-    if colspan
-      content_tag(:tr, class: "MatrixLine") do
-        content_tag(:td, tag.hr, class: "MatrixLine", colspan: colspan)
-      end
-    else
-      safe_empty
-    end
   end
 
   # Take URL that got us to this page and add one or more parameters to it.
@@ -418,28 +238,5 @@ module ApplicationHelper
     return nil unless idx
 
     query.result_ids[idx + 1] || query.result_ids[idx - 1]
-  end
-
-  # Override Rails method of the same name.  Just calls our
-  # Textile#textilize_without_paragraph method on the given string.
-  def textilize_without_paragraph(str, do_object_links = false)
-    Textile.textilize_without_paragraph(str, do_object_links)
-  end
-
-  # Override Rails method of the same name.  Just calls our Textile#textilize
-  # method on the given string.
-  def textilize(str, do_object_links = false)
-    Textile.textilize(str, do_object_links)
-  end
-
-  # contents of the <title> in html header
-  def title_tag_contents(action_name)
-    if @title.present?
-      @title.strip_html.html_safe
-    elsif TranslationString.where(tag: "title_for_#{action_name}").present?
-      :"title_for_#{action_name}".t
-    else
-      action_name.tr("_", " ").titleize
-    end
   end
 end
