+++ app/helpers/content_helper.rb
@@ -2,6 +2,75 @@
 
 # helpers for content tags
 module ContentHelper
+  def safe_empty
+    "".html_safe
+  end
+
+  def safe_br
+    "<br/>".html_safe
+  end
+
+  def safe_nbsp
+    "&nbsp;".html_safe
+  end
+
+  # Create an in-line white-space element approximately the given width in
+  # pixels.  It should be non-line-breakable, too.
+  def indent
+    content_tag(:span, "&nbsp;".html_safe, class: "ml-10px")
+  end
+
+  def spacer
+    content_tag(:span, "&nbsp;".html_safe, class: "mx-2")
+  end
+
+  # Return escaped HTML.
+  #
+  #   "<i>X</i>"  -->  "&lt;i&gt;X&lt;/i&gt;"
+  def escape_html(html)
+    h(html.to_str)
+  end
+
+  # Override Rails method of the same name.  Just calls our
+  # Textile#textilize_without_paragraph method on the given string.
+  def textilize_without_paragraph(str, do_object_links = false)
+    Textile.textilize_without_paragraph(str, do_object_links)
+  end
+
+  # Override Rails method of the same name.  Just calls our Textile#textilize
+  # method on the given string.
+  def textilize(str, do_object_links = false)
+    Textile.textilize(str, do_object_links)
+  end
+
+  # ----------------------------------------------------------------------------
+
+  def safe_spinner(text = "")
+    "#{text}<span class='spinner-right mx-2'></span>".html_safe
+  end
+
+  def content_tag_if(condition, name, content_or_options_with_block = nil,
+                     options = nil, escape = true, &block)
+    return unless condition
+
+    content_tag(name, content_or_options_with_block, options, escape, &block)
+  end
+
+  def content_tag_unless(condition, name, content_or_options_with_block = nil,
+                         options = nil, escape = true, &block)
+    content_tag_if(!condition, name, content_or_options_with_block,
+                   options, escape, &block)
+  end
+
+  # Wrap an html object in '<span title="blah">' tag.  This has the effect of
+  # giving it context help (mouse-over popup) in most modern browsers.
+  #
+  #   <%= add_context_help(link, "Click here to do something.") %>
+  #
+  def add_context_help(object, help)
+    content_tag(:span, object, title: help, data: { toggle: "tooltip" })
+  end
+
   # make a help-note styled element, like a div, p, or span
   def help_note(element = :span, string = "")
     content_tag(element, string, class: "help-note mr-3")
